zephyr-rust-docker: Do not point to latest
When I created this image, the upstream zephyr docker image has not been
tagged yet so I needed the latest toolchains.  However, now when I
build, I am pulling in stuff that is too new.  Point to the tagged
upstream of v0.13.1.  This means the toolchain will be two minor
revisions older, but I can live with that.